+++ src/Zerbikat/BackendBundle/Controller/AraumotaController.php
@@ -132,16 +132,24 @@ class AraumotaController extends Controller
      */
     public function deleteAction(Request $request, Araumota $araumotum)
     {
-        $form = $this->createDeleteForm($araumotum);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->remove($araumotum);
-            $em->flush();
+        $auth_checker = $this->get('security.authorization_checker');
+        if((($auth_checker->isGranted('ROLE_ADMIN')) && ($araumotum->getUdala()==$this->getUser()->getUdala()))
+            ||($auth_checker->isGranted('ROLE_SUPER_ADMIN')))
+        {
+            $form = $this->createDeleteForm($araumotum);
+            $form->handleRequest($request);
+    
+            if ($form->isSubmitted() && $form->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->remove($araumotum);
+                $em->flush();
+            }
+            return $this->redirectToRoute('araumota_index');
+        }else
+        {
+            //baimenik ez
+            return $this->redirectToRoute('fitxa_index');
         }
-
-        return $this->redirectToRoute('araumota_index');
     }
 
     /**
