+++ src/Zerbikat/BackendBundle/Controller/ArruntaController.php
@@ -134,16 +134,24 @@ class ArruntaController extends Controller
      */
     public function deleteAction(Request $request, Arrunta $arruntum)
     {
-        $form = $this->createDeleteForm($arruntum);
-        $form->handleRequest($request);
+        $auth_checker = $this->get('security.authorization_checker');
+        if((($auth_checker->isGranted('ROLE_ADMIN')) && ($arruntum->getUdala()==$this->getUser()->getUdala()))
+            ||($auth_checker->isGranted('ROLE_SUPER_ADMIN')))
+        {
+            $form = $this->createDeleteForm($arruntum);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->remove($arruntum);
-            $em->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->remove($arruntum);
+                $em->flush();
+            }
+            return $this->redirectToRoute('arrunta_index');
+        }else
+        {
+            //baimenik ez
+            return $this->redirectToRoute('fitxa_index');
         }
-
-        return $this->redirectToRoute('arrunta_index');
     }
 
     /**
