+++ src/Zerbikat/BackendBundle/Controller/AtalaController.php
@@ -135,16 +135,24 @@ class AtalaController extends Controller
      */
     public function deleteAction(Request $request, Atala $atala)
     {
-        $form = $this->createDeleteForm($atala);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->remove($atala);
-            $em->flush();
+        $auth_checker = $this->get('security.authorization_checker');
+        if((($auth_checker->isGranted('ROLE_ADMIN')) && ($atala->getUdala()==$this->getUser()->getUdala()))
+            ||($auth_checker->isGranted('ROLE_SUPER_ADMIN')))
+        {
+            $form = $this->createDeleteForm($atala);
+            $form->handleRequest($request);
+    
+            if ($form->isSubmitted() && $form->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->remove($atala);
+                $em->flush();
+            }
+            return $this->redirectToRoute('atala_index');
+        }else
+        {
+            //baimenik ez
+            return $this->redirectToRoute('fitxa_index');
         }
-
-        return $this->redirectToRoute('atala_index');
     }
 
     /**
