+++ src/Zerbikat/BackendBundle/Controller/BaldintzaController.php
@@ -134,16 +134,23 @@ class BaldintzaController extends Controller
      */
     public function deleteAction(Request $request, Baldintza $baldintza)
     {
-        $form = $this->createDeleteForm($baldintza);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->remove($baldintza);
-            $em->flush();
+        $auth_checker = $this->get('security.authorization_checker');
+        if((($auth_checker->isGranted('ROLE_ADMIN')) && ($baldintza->getUdala()==$this->getUser()->getUdala()))
+            ||($auth_checker->isGranted('ROLE_SUPER_ADMIN')))
+        {
+            $form = $this->createDeleteForm($baldintza);
+            $form->handleRequest($request);
+            if ($form->isSubmitted() && $form->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->remove($baldintza);
+                $em->flush();
+            }
+            return $this->redirectToRoute('baldintza_index');
+        }else
+        {
+            //baimenik ez
+            return $this->redirectToRoute('fitxa_index');
         }
-
-        return $this->redirectToRoute('baldintza_index');
     }
 
     /**
