+++ src/Zerbikat/BackendBundle/Controller/DoklagunController.php
@@ -133,16 +133,25 @@ class DoklagunController extends Controller
      */
     public function deleteAction(Request $request, Doklagun $doklagun)
     {
-        $form = $this->createDeleteForm($doklagun);
-        $form->handleRequest($request);
+        $auth_checker = $this->get('security.authorization_checker');
+        if((($auth_checker->isGranted('ROLE_ADMIN')) && ($doklagun->getUdala()==$this->getUser()->getUdala()))
+            ||($auth_checker->isGranted('ROLE_SUPER_ADMIN')))
+        {
+            $form = $this->createDeleteForm($doklagun);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->remove($doklagun);
-            $em->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->remove($doklagun);
+                $em->flush();
+            }
+
+            return $this->redirectToRoute('doklagun_index');
+        }else
+        {
+            //baimenik ez
+            return $this->redirectToRoute('fitxa_index');
         }
-
-        return $this->redirectToRoute('doklagun_index');
     }
 
     /**
