+++ src/Zerbikat/BackendBundle/Controller/EremuakController.php
@@ -133,16 +133,22 @@ class EremuakController extends Controller
      */
     public function deleteAction(Request $request, Eremuak $eremuak)
     {
-        $form = $this->createDeleteForm($eremuak);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->remove($eremuak);
-            $em->flush();
+        $auth_checker = $this->get('security.authorization_checker');
+        if($auth_checker->isGranted('ROLE_SUPER_ADMIN'))
+        {
+            $form = $this->createDeleteForm($eremuak);
+            $form->handleRequest($request);
+            if ($form->isSubmitted() && $form->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->remove($eremuak);
+                $em->flush();
+            }
+            return $this->redirectToRoute('eremuak_index');
+        }else
+        {
+            //baimenik ez
+            return $this->redirectToRoute('fitxa_index');
         }
-
-        return $this->redirectToRoute('eremuak_index');
     }
 
     /**
