+++ src/Zerbikat/BackendBundle/Controller/IsiltasunAdministratiboaController.php
@@ -132,16 +132,23 @@ class IsiltasunAdministratiboaController extends Controller
      */
     public function deleteAction(Request $request, IsiltasunAdministratiboa $isiltasunAdministratiboa)
     {
-        $form = $this->createDeleteForm($isiltasunAdministratiboa);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->remove($isiltasunAdministratiboa);
-            $em->flush();
+        $auth_checker = $this->get('security.authorization_checker');
+        if((($auth_checker->isGranted('ROLE_ADMIN')) && ($isiltasunAdministratiboa->getUdala()==$this->getUser()->getUdala()))
+            ||($auth_checker->isGranted('ROLE_SUPER_ADMIN')))
+        {
+            $form = $this->createDeleteForm($isiltasunAdministratiboa);
+            $form->handleRequest($request);
+            if ($form->isSubmitted() && $form->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->remove($isiltasunAdministratiboa);
+                $em->flush();
+            }
+            return $this->redirectToRoute('isiltasunadministratiboa_index');
+        }else
+        {
+            //baimenik ez
+            return $this->redirectToRoute('fitxa_index');
         }
-
-        return $this->redirectToRoute('isiltasunadministratiboa_index');
     }
 
     /**
